refactor: deduplicate cluster fixture
