+++ controllers/maskinportenclient/maskinportenclient_controller_test.go
@@ -7,7 +7,6 @@ import (
 	v1 "github.com/nais/digdirator/api/v1"
 	"github.com/nais/digdirator/pkg/config"
 	"github.com/nais/digdirator/pkg/fixtures"
-	"github.com/nais/digdirator/pkg/fixtures/maskinporten"
 	"github.com/nais/digdirator/pkg/labels"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
@@ -137,7 +136,7 @@ func TestMaskinportenController(t *testing.T) {
 	}
 
 	// set up preconditions for cluster
-	clusterFixtures := maskinporten.New(cli, cfg).MinimalConfig().WithPod().WithUnusedSecret()
+	clusterFixtures := fixtures.New(cli, cfg).MinimalConfig().WithMaskinportenClient().WithPod().WithUnusedSecret()
 
 	// create MaskinportenClient
 	if err := clusterFixtures.Setup(); err != nil {
